+++ tests/eintr.py
@@ -5,7 +5,6 @@ check issue #8
 from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
-from __future__ import unicode_literals
 
 __all__ = ['main']